fix(web): preserve snowflake IDs through JSON number decoding.
Flutter Web was rounding int64 post IDs from the recommend feed, so
opening 继续联调帖 requested a nearby ID and the gateway returned 内容不存在.
Decode 16+ digit JSON integers as strings, keep those digits in routes
and request paths, and re-emit JSON numbers for the Go int64 contract.

+++ lib/features/profile/presentation/edit_profile_page.dart
@@ -43,7 +43,7 @@ class _EditProfilePageState extends ConsumerState<EditProfilePage> {
     try {
       final user = await ref
           .read(_userRepoProvider)
-          .getUserProfile(auth.userId!.toInt());
+          .getUserProfile(auth.userId!);
       setState(() {
         _nicknameCtrl.text = user.nickname;
         _bioCtrl.text = user.bio;
